batch/requirements.txt
python-dotenv
openai
pyupbit
ta
Pillow
mysql-connector-python
cryptography
boto3
schedule
# pydantic-settings
# selenium
# webdriver-manager